pyproject.toml
[tool.poetry]
name = "analyst"
version = "0.1.0"
description = "A personal, experimental financial analysis tool."
authors = ["sfjt"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.4.1"
python-dotenv = "^1.0.0"
polygon-api-client = "^1.12.3"
flask = "^2.3.2"
flask-pymongo = "^2.3.0"
pandas = "^2.0.3"
mplfinance = "^0.12.10b0"
pytest-mock = "^3.11.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.7.0"
pytest = "^7.4.0"
pytest-dotenv = "^0.5.2"
mongomock = "^4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = ".venv"
